pythonesec(not-yet-combine)/requirements.txt

Flask

yara-python
Pillow
pyzbar
opencv-python
numpy